move bufferCounter into utils
