+++ lib/utils.ts
@@ -1,5 +1,11 @@
 import { createHash } from 'crypto';
 
+import { complement, forEach, isEmpty, tap, times } from 'ramda';
+
+import { bond } from 'proxy-bind';
+
+import { Read } from 'async-readable';
+
 
 
 
@@ -76,5 +82,80 @@ export function blockHash (content: Buffer) {
 
 export function reverseBuffer (buffer: Buffer) {
     return buffer.reverse() as Buffer;
+}
+
+
+
+export function bufferCounter (read: Read) {
+
+    const chunks = [] as Buffer[];
+    const marker = [] as number[];
+
+    let flag = false;
+
+    const notEmpty = complement(isEmpty);
+    const mirror = tap(bond(chunks).push);
+    const concatChunks = () => Buffer.concat(chunks);
+    const patchChunksBy = forEach(((x) => (i: number) => chunks[i] = x)(Buffer.alloc(0)));
+    const markChunksFromBack = (offset: number) => marker.push(chunks.length - 1 - offset);
+
+
+
+    return Object.freeze({
+
+        flag (on: boolean) {
+            flag = on;
+        },
+
+        async read (size: number) {
+
+            const chunk = mirror(await read(size));
+
+            if (flag === true) {
+                markChunksFromBack(0);
+            }
+
+            return chunk;
+
+        },
+
+        pop (n: number) {
+            times(markChunksFromBack, n);
+        },
+
+        reset () {
+            chunks.length = 0;
+            marker.length = 0;
+            flag = false;
+        },
+
+        count () {
+
+            const total = concatChunks();
+            const totalBytes = total.length;
+
+            let general = total;
+            let generalBytes = totalBytes;
+
+            if (notEmpty(marker)) {
+                patchChunksBy(marker);
+
+                general = concatChunks();
+                generalBytes = general.length;
+            }
+
+            const weight = generalBytes * 3 + totalBytes;
+            const hash = blockHash(general);
+
+            return {
+                weight,
+                hash,
+                size: totalBytes,
+            };
+
+        },
+
+    });
+
 }
